alx_backend_security/requirements.txt
amqp==5.3.1
anyio==4.10.0
asgiref==3.9.0
backoff==2.2.1
billiard==4.2.1
blinker==1.9.0
celery==5.5.3
certifi==2025.8.3
chapa-cli==0.0.4
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.5
Django==5.2.4
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-crontab==0.7.1
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-ipware==7.0.1
django-redis==6.0.0
django-seed==0.3.1
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.94.2
drf-yasg==1.21.10
Faker==37.4.2
Flask==3.1.1
future==1.0.0
git-filter-repo==2.47.0
gql==4.0.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
mariadb==1.1.12
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.6.4
mysqlclient==2.2.7
packaging==25.0
parameterized==0.9.0
promise==2.3
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
pyngrok==7.3.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-ipware==3.0.0
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
serializers==0.2.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
text-unidecode==1.3
toposort==1.10
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.20.1
